requirements.txt
# 金融交易系统依赖
# ==================

# 核心科学计算
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0

# GPU 加速
cupy-cuda12x>=12.3.0
numba>=0.58.0

# 金融数据
alpha-vantage>=2.3.0
yfinance>=0.2.30
requests>=2.31.0
aiohttp>=3.9.0

# 技术分析
ta-lib-bin>=0.4.28
pandas-ta>=0.3.14b0

# 数据库
sqlalchemy>=2.0.20
alembic>=1.12.0
aiosqlite>=0.19.0
redis>=5.0.0
h5py>=3.10.0

# 可视化
matplotlib>=3.8.0
plotly>=5.18.0
seaborn>=0.13.0

# 配置和日志
pyyaml>=6.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
rich>=13.7.0
typer>=0.9.0

# 异步和并发
aiohttp>=3.9.0
joblib>=1.3.0

# 开发工具
jupyterlab>=4.0.0
ipywidgets>=8.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
isort>=5.12.0
mypy>=1.7.0
ruff>=0.1.0

# 机器学习（后续模块）
scikit-learn>=1.3.0
torch>=2.1.0

# 类型检查
typing-extensions>=4.8.0
pydantic>=2.5.0
